fixup! FM: add the ability to transmit DTMF tones on some platforms. Remap 1750 Hz access tone to F1 key.
