+++ platform/drivers/baseband/radio_UV3x0.cpp
@@ -29,7 +29,6 @@
 #include "radioUtils.h"
 #include "drivers/baseband/HR_C6000.h"
 #include "drivers/baseband/AT1846S.h"
-#include "protocols/FM/DTMF.hpp"
 
 static const rtxStatus_t *config;                // Pointer to data structure with radio configuration
 
@@ -259,7 +258,7 @@ void radio_enableTx()
 
     if (config->dtmfPressed != 0)
     {
-        DtmfKeyEntry entry = DTMF::lookupDtmfFreq(config->dtmfPressed);
+        DtmfKeyEntry entry = lookupDtmfFreq(config->dtmfPressed);
         const uint16_t scaleFactor = 10; // Convert Hz to AT1846S frequency units (0.1 Hz)
         if (entry.symbol != '\0') {
             at1846s.enableDtmfTone(entry.freqLow * scaleFactor, entry.freqHigh * scaleFactor);
