+++ scripts/clang_format.sh
@@ -69,10 +69,8 @@ openrtx/include/interfaces/radio.h
 openrtx/include/peripherals/gps.h
 openrtx/include/peripherals/rng.h
 openrtx/include/peripherals/rtc.h
-openrtx/include/protocols/FM/DTMF.hpp
 openrtx/src/core/dsp.cpp
 openrtx/src/core/memory_profiling.cpp
-openrtx/src/protocols/FM/DTMF.cpp
 platform/drivers/ADC/ADC0_GDx.h
 platform/drivers/audio/MAX9814.h
 platform/drivers/baseband/MCP4551.h
